+++ packages/browser-wallet/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 -   Removed EUROe as default token and associated migrations with EUROe as default token. EUROe can still be added by the "Manage token list" menu manually.
+-   Added a fix where the user was not able to do web3id age verification
 
 ## 2.5.0